Initial working version
Current features:
- Any file writes changes the client view to that of the new file
- Writes to the same file will automatically scroll to the bottom
- Recursive directory watching

The current version should be usable for notetaking. I will be
implementing some UI upgrades and QOL changes, such as scrolling to
the first position that is different in sample file writes. Also,
linking between files is not supported in page as any links will open a
new tab or redirect.